requirements.txt
asgiref
autopep8
django==3.2
pycodestyle
pytz
sqlparse
Unipath~=1.1
dj-database-url
python-decouple
gunicorn
whitenoise
psycopg2-binary
django-filter~=23.5
djangorestframework~=3.15.1
django-debug-toolbar
django-colorfield
matplotlib
pandas~=1.3.5
django-mptt~=0.14.0
django-admin-tools~=0.9.3
django-admin-charts
django-import-export
django-admin-footer
data-wizard
openai~=1.17.0
opencv-python~=4.9.0.80
moviepy~=1.0.3
psycopg2~=2.9.9
requests~=2.27.1
numpy~=1.21.6
scikit-learn~=1.0.2
keras~=2.11.0
mtcnn~=0.1.1
Pillow~=9.5.0